backend/requirements.txt
anyio==3.3.4
asgiref==3.4.1
autopep8==1.6.0
backports.entry-points-selectable==1.1.1
bcrypt==3.2.0
CacheControl==0.12.10
cachy==0.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cleo==0.8.1
click==8.0.3
clikit==0.6.2
crashtest==0.3.1
cryptography==35.0.0
distlib==0.3.3
ecdsa==0.17.0
fastapi==0.70.0
filelock==3.3.2
h11==0.12.0
html5lib==1.1
idna==3.3
jeepney==0.7.1
keyring==21.8.0
lockfile==0.12.2
msgpack==1.0.2
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
phonenumbers==8.12.37
pkginfo==1.7.1
platformdirs==2.4.0
poetry==1.1.11
poetry-core==1.0.7
psycopg2-binary==2.9.2
ptyprocess==0.7.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
pylev==1.4.0
pyparsing==3.0.6
python-decouple==3.5
python-jose==3.3.0
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.7.2
SecretStorage==3.3.1
shellingham==1.4.0
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
toml==0.10.2
tomlkit==0.7.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
virtualenv==20.10.0
webencodings==0.5.1
